+++ pczt/src/roles/tx_extractor/mod.rs
@@ -10,9 +10,12 @@ use zcash_primitives::transaction::{
     txid::TxIdDigester,
     Authorization, Transaction, TransactionData, TxVersion,
 };
-use zcash_protocol::consensus::BranchId;
+use zcash_protocol::{
+    consensus::BranchId,
+    constants::{V5_TX_VERSION, V5_VERSION_GROUP_ID},
+};
 
-use crate::{common::determine_lock_time, Pczt, V5_TX_VERSION, V5_VERSION_GROUP_ID};
+use crate::{common::determine_lock_time, Pczt};
 
 mod orchard;
 pub use self::orchard::OrchardError;
@@ -78,7 +81,7 @@ impl<'a> TransactionExtractor<'a> {
         } = self;
 
         let version = match (pczt.global.tx_version, pczt.global.version_group_id) {
-            (V5_TX_VERSION, V5_VERSION_GROUP_ID) => Ok(TxVersion::Zip225),
+            (V5_TX_VERSION, V5_VERSION_GROUP_ID) => Ok(TxVersion::V5),
             (version, version_group_id) => Err(Error::Global(GlobalError::UnsupportedTxVersion {
                 version,
                 version_group_id,